# public/swagger.json
{
	"components": {
...
		"schemas": {
			"mongoose.Types.ObjectId": {
				"type": "string"
			},
			"ICollege": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"provider": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"name": {
						"properties": {
							"middleName": {
								"type": "string"
							},
							"givenName": {
								"type": "string"
							},
							"familyName": {
								"type": "string"
							}
						},
						"required": [
							"givenName",
							"familyName"
						],
						"type": "object"
					},
					"email": {
						"type": "string"
					},
					"emails": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"photos": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"college": {
						"$ref": "#/components/schemas/ICollege"
					}
				},
				"required": [
					"_id",
					"id",
					"provider",
					"displayName",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ICategory": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ICategory.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ICreateCategory": {
				"$ref": "#/components/schemas/Pick_ICategory.name_"
			},
			"Image": {
				"properties": {
					"url": {
						"type": "string"
					}
				},
				"required": [
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProduct": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"name": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"gallery": {
						"items": {
							"$ref": "#/components/schemas/Image"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"condition": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"category": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"thumbnail",
					"gallery",
					"price",
					"condition",
					"category",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IMessage": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"chat": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"sender": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"text": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"chat",
					"sender",
					"text"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IChatDTO": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"buyer": {
						"$ref": "#/components/schemas/IUser"
					},
					"product": {
						"$ref": "#/components/schemas/IProduct"
					},
					"acceptedPrice": {
						"type": "number",
						"format": "double"
					},
					"owner": {
						"$ref": "#/components/schemas/IUser"
					},
					"messages": {
						"items": {
							"$ref": "#/components/schemas/IMessage"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"buyer",
					"product",
					"owner"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IChat": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"buyer": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"owner": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"product": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"acceptedPrice": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"buyer",
					"owner",
					"product",
					"acceptedPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IChat.buyer-or-product_": {
				"properties": {
					"buyer": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"product": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					}
				},
				"required": [
					"buyer",
					"product"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ICreateChat": {
				"$ref": "#/components/schemas/Pick_IChat.buyer-or-product_"
			},
...
			"Pick_IMessage.chat-or-text_": {
				"properties": {
					"chat": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"text": {
						"type": "string"
					}
				},
				"required": [
					"chat",
					"text"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ICreateMessage": {
				"$ref": "#/components/schemas/Pick_IMessage.chat-or-text_"
			},
...
			"IProductResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"description": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"gallery": {
						"items": {
							"$ref": "#/components/schemas/Image"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"condition": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"college": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"owner": {
						"$ref": "#/components/schemas/IUser"
					}
				},
				"required": [
					"name",
					"_id",
					"description",
					"thumbnail",
					"gallery",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProductFilter": {
				"properties": {
					"categories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"colleges": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"categories",
					"colleges"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISingleProductResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"description": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"gallery": {
						"items": {
							"$ref": "#/components/schemas/Image"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"condition": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"college": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"owner": {
						"$ref": "#/components/schemas/IUser"
					},
					"similar": {
						"items": {
							"$ref": "#/components/schemas/IProductResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"_id",
					"description",
					"thumbnail",
					"gallery",
					"price",
					"similar"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IProduct.Exclude_keyofIProduct._id__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"condition": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"category": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"thumbnail": {
						"type": "string"
					},
					"gallery": {
						"items": {
							"$ref": "#/components/schemas/Image"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"description",
					"condition",
					"category",
					"thumbnail",
					"gallery",
					"price"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IProduct._id_": {
				"$ref": "#/components/schemas/Pick_IProduct.Exclude_keyofIProduct._id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"ProductDto": {
				"$ref": "#/components/schemas/Omit_IProduct._id_"
			},
...
		},
		"securitySchemes": {
			"jwtAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Bearer token to access these API endpoints"
			}
		}
	},
	"info": {
		"title": "commerce-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/categories": {
			"get": {
				"operationId": "GetAllCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ICategory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICategory"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateCategory"
							}
						}
					}
				}
			}
		},
		"/api/categories/{categoryId}": {
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICategory"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateCategory"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/chats/{chatId}/by-id": {
			"get": {
				"operationId": "GetChatById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IChatDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/chats/my-chats": {
			"get": {
				"operationId": "MyChats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IChatDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/chats/{productId}": {
			"get": {
				"operationId": "GetChat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IChatDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/chats/{chatId}": {
			"put": {
				"operationId": "UpdateChat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IChat"
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateChat"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteChat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/messages": {
			"post": {
				"operationId": "CreateMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessage"
								}
							}
						}
					}
				},
				"tags": [
					"Messages"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateMessage"
							}
						}
					}
				}
			}
		},
		"/api/messages/{messageId}": {
			"put": {
				"operationId": "UpdateMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessage"
								}
							}
						}
					}
				},
				"tags": [
					"Messages"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateMessage"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Messages"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IProductResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProduct"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductDto"
							}
						}
					}
				}
			}
		},
		"/api/products/filter": {
			"post": {
				"operationId": "FilterProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProductFilter"
							}
						}
					}
				}
			}
		},
		"/api/products/{productId}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISingleProductResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": []
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProduct"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductDto"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}